refactor: CartRepository 추상화
