+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -5,7 +5,7 @@ import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
-import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.CartDefaultRepository
 import woowacourse.shopping.databinding.ActivityCartBinding
 
 class CartActivity : AppCompatActivity() {
@@ -13,7 +13,7 @@ class CartActivity : AppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
 
     private val viewModel: CartViewModel by viewModels {
-        CartViewModel.CartViewModelFactory(CartRepository())
+        CartViewModel.CartViewModelFactory(CartDefaultRepository())
     }
 
     private lateinit var dateFormatter: DateFormatter
