+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.repository.CartRepository
 
 class CartViewModel(
     private val cartRepository: CartRepository,
