+++ domain/src/main/java/woowacourse/shopping/repository/CartRepository.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.repository
+
+import woowacourse.shopping.model.Product
+
+interface CartRepository {
+
+    fun addCartProduct(product: Product)
+
+    fun getAllCartProducts(): List<Product>
+
+    fun deleteCartProduct(id: Int)
+}
